replace the old rust implementation with new
