--- rust-usfm-parser-new/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "usfm-grammar"
-version = "3.0.0-alpha.1"
-edition = "2024"
-
-[dependencies]
-tree-sitter        = "0.26.8"
-tree-sitter-usfm3  = "3.0.0-alpha.2"
-serde_json         = "1"
-elementtree        = "1.2"
-regex              = "1" 
-streaming-iterator = "0.1"
-clap = { version = "4", features = ["derive"] }
-
-[dev-dependencies]
-glob     = "0.3"
