--- rust-usfm-parser-new/src/lib.rs
@@ -1,10 +1,0 @@
-pub mod usfm_parser;
-
-mod list_generator;
-mod usj_generator;
-mod usfm_generator;
-mod usx_generator;
-mod filter;
-
-pub use usfm_parser::{USFMParser, Filter, Format};
-pub use usfm_generator::BibleNlpInput;
