+++ rust-usfm-parser/Cargo.toml
@@ -1,33 +1,16 @@
 [package]
-name = "rust-usfm"
-version = "0.1.0"
-edition = "2021"
-
-
-
-
-
-[build-dependencies]
-cc="*"
-
+name = "usfm-grammar"
+version = "3.0.0-alpha.1"
+edition = "2024"
 
 [dependencies]
-tree-sitter = "0.24.7"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-jsonschema = "0.16"
-tree-sitter-rust = "0.21"
-tree-sitter-usfm3 = "3.0.0-alpha.1"
-regex = "1"
-lazy_static = "1.4"
-once_cell = "1.18"
-structopt = "0.3"
-streaming-iterator = "0.1.9"
-glob = "0.3"
-strum = "0.24"
-strum_macros = "0.24"
-quick-xml = "0.23"
-roxmltree = "0.19"
-rayon = "1.5"
-log = "0.4"
-env_logger = "0.10"
+tree-sitter        = "0.26.8"
+tree-sitter-usfm3  = "3.0.0-alpha.2"
+serde_json         = "1"
+elementtree        = "1.2"
+regex              = "1" 
+streaming-iterator = "0.1"
+clap = { version = "4", features = ["derive"] }
+
+[dev-dependencies]
+glob     = "0.3"
