+++ rust-usfm-parser/src/lib.rs
@@ -1,5 +1,10 @@
-// src/lib.rs
-//test comment
-pub mod globals;
-pub mod parser;
-pub mod usj_generator;
+pub mod usfm_parser;
+
+mod list_generator;
+mod usj_generator;
+mod usfm_generator;
+mod usx_generator;
+mod filter;
+
+pub use usfm_parser::{USFMParser, Filter, Format};
+pub use usfm_generator::BibleNlpInput;
