+++ rust-usfm-parser/src/list_generator.rs
