+++ rust-usfm-parser/src/usj_generator.rs
@@ -1,1204 +1,848 @@
-use crate::globals::GLOBAL_TREE;
-use regex::Regex;
+use serde_json::{json, Value};
+use tree_sitter::Node;
 
-extern crate lazy_static;
+// ---------------------------------------------------------------------------
+// Marker constants – mirrors USXGenerator / USJGenerator.MARKER_LISTS
+// ---------------------------------------------------------------------------
 
-use lazy_static::lazy_static;
-use serde_json::{self, json};
-use std::collections::HashMap;
-use std::sync::Mutex;
-use std::sync::MutexGuard;
-use streaming_iterator::StreamingIterator;
-use tree_sitter::{Query, QueryCursor, TextProvider};
-use tree_sitter_usfm3;
-const NOTE_MARKERS: [&str; 6] = ["f", "fe", "ef", "efe", "x", "ex"];
-const CHAR_STYLE_MARKERS: [&str; 55] = [
-    "add", "bk", "dc", "ior", "iqt", "k", "litl", "nd", "ord", "pn", "png", "qac", "qs", "qt",
-    "rq", "sig", "sls", "tl", "wj", "em", "bd", "bdit", "it", "no", "sc", "sup", "rb", "pro", "w",
-    "wh", "wa", "wg", "lik", "liv", "jmp", "fr", "ft", "fk", "fq", "fqa", "fl", "fw", "fp", "fv",
-    "fdc", "xo", "xop", "xt", "xta", "xk", "xq", "xot", "xnt", "xdc", "ref",
-];
-const NESTED_CHAR_STYLE_MARKERS: [&str; 55] = [
-    "addNested",
-    "bkNested",
-    "dcNested",
-    "iorNested",
-    "iqtNested",
-    "kNested",
-    "litlNested",
-    "ndNested",
-    "ordNested",
-    "pnNested",
-    "pngNested",
-    "qacNested",
-    "qsNested",
-    "qtNested",
-    "rqNested",
-    "sigNested",
-    "slsNested",
-    "tlNested",
-    "wjNested",
-    "emNested",
-    "bdNested",
-    "bditNested",
-    "itNested",
-    "noNested",
-    "scNested",
-    "supNested",
-    "rbNested",
-    "proNested",
-    "wNested",
-    "whNested",
-    "waNested",
-    "wgNested",
-    "likNested",
-    "livNested",
-    "jmpNested",
-    "frNested",
-    "ftNested",
-    "fkNested",
-    "fqNested",
-    "fqaNested",
-    "flNested",
-    "fwNested",
-    "fpNested",
-    "fvNested",
-    "fdcNested",
-    "xoNested",
-    "xopNested",
-    "xtNested",
-    "xtaNested",
-    "xkNested",
-    "xqNested",
-    "xotNested",
-    "xntNested",
-    "xdcNested",
-    "refNested",
-];
-const PARA_STYLE_MARKERS: [&str; 49] = [
-    "ide", "h", "toc", "toca", //identification
-    "imt", "is", "ip", "ipi", "im", "imi", "ipq", "imq", "ipr", "iq", "ib", "ili", "iot", "io",
-    "iex", "imte", "ie", // intro
-    "mt", "mte", "cl", "cd", "ms", "mr", "s", "sr", "r", "d", "sp", "sd", //titles
-    "q", "qr", "qc", "qa", "qm", "qd", //poetry
-    "lh", "li", "lf", "lim", "litl", //lists
-    "sts", "rem", "lit", "restore", //comments
+const PARA_STYLE_MARKERS: &[&str] = &[
+    // identification
+    "ide", "h", "toc", "toca",
+    // intro
+    "imt", "is", "ip", "ipi", "im", "imi", "ipq", "imq",
+    "ipr", "iq", "ib", "ili", "iot", "io", "iex", "imte", "ie",
+    // titles
+    "mt", "mte", "cl", "cd", "ms", "mr", "s", "sr", "r", "d", "sp", "sd",
+    // poetry
+    "q", "qr", "qc", "qa", "qm", "qd",
+    // lists
+    "lh", "li", "lf", "lim",
+    // comments
+    "sts", "rem", "lit", "restore",
     "b",
 ];
-const DEFAULT_ATTRIB_MAP: [(&str, &str); 9] = [
-    ("w", "lemma"),
-    ("rb", "gloss"),
-    ("xt", "href"),
-    ("fig", "alt"),
-    ("xt_standalone", "href"),
-    ("xtNested", "href"),
-    ("ref", "loc"),
-    ("milestone", "who"),
-    ("k", "key"),
+
+const NOTE_MARKERS: &[&str] = &["f", "fe", "ef", "efe", "x", "ex"];
+
+const CHAR_STYLE_MARKERS: &[&str] = &[
+    // special-text
+    "add", "bk", "dc", "ior", "iqt", "k", "litl", "nd", "ord",
+    "pn", "png", "qac", "qs", "qt", "rq", "sig", "sls", "tl", "wj",
+    // character styling
+    "em", "bd", "bdit", "it", "no", "sc", "sup",
+    // special-features
+    "rb", "pro", "w", "wh", "wa", "wg",
+    // structured list entries
+    "lik", "liv",
+    // footnote-content
+    "jmp", "fr", "ft", "fk", "fq", "fqa", "fl", "fw", "fp", "fv", "fdc",
+    // crossref-content
+    "xo", "xop", "xt", "xta", "xk", "xq", "xot", "xnt", "xdc",
 ];
-const TABLE_CELL_MARKERS: [&str; 4] = ["tc", "th", "tcr", "thr"];
-lazy_static! {
-    static ref CHAPTER_NUMBER: Mutex<Option<String>> = Mutex::new(None);
+
+const TABLE_CELL_MARKERS: &[&str] = &["tc", "th", "tcr", "thr", "tcc", "thc"];
+
+const OTHER_PARA_NESTABLES: &[&str] = &[
+    "text", "footnote", "crossref", "verseText", "v", "b", "milestone", "zNameSpace",
+];
+
+/// Default attribute name when the attribute name node is "|"
+fn default_attrib_map(node_type: &str) -> Option<&'static str> {
+    match node_type {
+        "w"            => Some("lemma"),
+        "rb"           => Some("gloss"),
+        "xt"           => Some("href"),
+        "fig"          => Some("alt"),
+        "xt_standalone" => Some("href"),
+        "xtNested"     => Some("href"),
+        "ref"          => Some("loc"),
+        "milestone"    => Some("who"),
+        "k"            => Some("key"),
+        _              => None,
+    }
 }
 
-/*const NESTED_CHAR_STYLE_MARKERS: Vec<String> = CHAR_STYLE_MARKERS
-    .iter()
-    .map(|item| format!("{}Nested", item)) // Append "Nested" to each item
-    .collect();
-
-
-const COMBINED_MARKERS: Vec<&str> = CHAR_STYLE_MARKERS.iter().chain(NESTED_CHAR_STYLE_MARKERS.iter()).chain(vec!["xt_standalone", "ref"].iter()).cloned().collect();*/
-pub fn usj_generator(usfm: &str) -> Result<String, Box<dyn std::error::Error>> {
-    let global_tree: MutexGuard<Option<tree_sitter::Tree>> = GLOBAL_TREE.lock().unwrap();
-    let tree = global_tree.as_ref().ok_or("Tree is not initialized")?;
-
-    let root_node = tree.root_node();
-
-    // Change the type of json_object to HashMap<String, serde_json::Value>
-    let mut json_object: HashMap<String, serde_json::Value> = HashMap::new();
-    json_object.insert("type".to_string(), json!("USJ")); // Use json! macro for consistency
-    json_object.insert("version".to_string(), json!("3.1")); // Use json! macro for consistency
-
-    let mut content = Vec::new();
-
-    // Traverse the tree and build the JSON object
-    node_2_usj(&root_node, &mut content, usfm);
-    // Insert the content as a serde_json::Value
-    json_object.insert(
-        "content".to_string(),
-        serde_json::to_value(content).unwrap(),
-    );
-
-    // Convert HashMap to pretty-printed JSON
-    let json_output = serde_json::to_string_pretty(&json_object).unwrap_or_else(|e| {
-        eprintln!("Failed to convert to JSON: {}", e);
-        String::new() // Return an empty string or handle the error as needed
-    });
-    Ok(json_output)
+fn is_nested_char(t: &str) -> bool {
+    CHAR_STYLE_MARKERS.iter().any(|m| format!("{}Nested", m) == t)
 }
-fn node_2_usj(node: &tree_sitter::Node, parent_json_obj: &mut Vec<serde_json::Value>, usfm: &str) {
-    println!("Node Type: {}", node.kind());
-    
-    if node.kind() == "id" {
-        node_2_usj_id(&node, parent_json_obj, usfm);
-    } else if node.kind() == "chapter" {
-        node_2_usj_chapter(&node, parent_json_obj, usfm);
-    } else if ["cl", "cp", "cd", "vp"].contains(&node.kind()) {
-        node_2_usj_generic(&node, parent_json_obj, usfm);
-    } else if ["ca", "va"].contains(&node.kind()) {
-        node_2_usj_ca_va(&node, parent_json_obj, usfm);
-    } else if node.kind() == "v" {
-        let mut global_chapter_number = CHAPTER_NUMBER.lock().unwrap();
-        node_2_usj_verse(&node, parent_json_obj, usfm,&global_chapter_number);    } else if node.kind() == "verseText" {
-        for child in node.children(&mut node.walk()) {
-            node_2_usj(&child, parent_json_obj, usfm);
+
+fn in_set(t: &str, set: &[&str]) -> bool {
+    set.contains(&t)
+}
+
+// ---------------------------------------------------------------------------
+// Parse state
+// ---------------------------------------------------------------------------
+
+#[derive(Default)]
+struct ParseState {
+    book_slug:       Option<String>,
+    current_chapter: Option<String>,
+}
+
+// ---------------------------------------------------------------------------
+// USJGenerator
+// ---------------------------------------------------------------------------
+
+pub struct USJGenerator<'a> {
+    usfm:        &'a [u8],
+    parse_state: ParseState,
+}
+
+impl<'a> USJGenerator<'a> {
+    pub fn new(usfm: &'a [u8]) -> Self {
+        Self {
+            usfm,
+            parse_state: ParseState::default(),
         }
-    } else if ["paragraph", "pi", "ph"].contains(&node.kind()) {
-        node_2_usj_para(&node, parent_json_obj, usfm);
-    } else if NOTE_MARKERS.contains(&node.kind()) {
-        node_2_usj_notes(&node, parent_json_obj, usfm);
-    } else if CHAR_STYLE_MARKERS.contains(&node.kind())
-        || NESTED_CHAR_STYLE_MARKERS.contains(&node.kind())
-        || node.kind() == "xt_standalone"
-    {
-        node_2_usj_char(&node, parent_json_obj, usfm);
-    } else if node.kind().ends_with("Attribute") {
-        node_2_usj_attrib(&node, parent_json_obj, usfm);
-    } else if node.kind() == "text" {
-        let node_text = node
-            .utf8_text(usfm.as_bytes())
-            .expect("Failed to get node text")
-            .replace('\n', "")
-            .to_string();
-        if node_text != "" {
-            parent_json_obj.push(json!(node_text));
+    }
+
+    // -----------------------------------------------------------------------
+    // Public entry point
+    // -----------------------------------------------------------------------
+
+    pub fn get_usj(&mut self, root: Node) -> Value {
+        let mut usj_root = json!({
+            "type":    "USJ",
+            "version": "3.1",
+            "content": []
+        });
+        self.node_2_usj(root, &mut usj_root);
+        usj_root
+    }
+
+    // -----------------------------------------------------------------------
+    // Helper: slice bytes from node
+    // -----------------------------------------------------------------------
+
+    fn node_text(&self, node: Node) -> String {
+        let bytes = &self.usfm[node.start_byte()..node.end_byte()];
+        String::from_utf8_lossy(bytes).into_owned()
+    }
+
+    fn node_text_trimmed(&self, node: Node) -> String {
+        self.node_text(node).trim().to_string()
+    }
+
+    // -----------------------------------------------------------------------
+    // Helper: push a string into a content array
+    // -----------------------------------------------------------------------
+
+    fn push_str(parent: &mut Value, s: String) {
+        if let Some(arr) = parent["content"].as_array_mut() {
+            arr.push(Value::String(s));
         }
-    }else if ["table", "tr"].contains(&node.kind()) {
-        node_2_usj_table(&node, parent_json_obj, usfm);
-    } else if "zNameSpace".contains(&node.kind()) {
-        node_2_usj_milestone(&node, parent_json_obj, usfm);
-    }else if "milestone".contains(&node.kind()) {
-        node_2_usj_milestone(&node, parent_json_obj, usfm);
-    } else if ["esb", "cat", "fig"].contains(&node.kind()) {
-        node_2_usj_special(&node, parent_json_obj, usfm);
-    } else if PARA_STYLE_MARKERS.contains(&node.kind())
-        || PARA_STYLE_MARKERS.contains(&node.kind().replace("\\", "").trim())
-    {
-        node_2_usj_generic(node, parent_json_obj, usfm);
-    } else if ["", "|"].contains(&node.kind().trim()) {
-        // skip white space nodes
-    } else if node.named_child_count() > 0 {
+    }
+
+    // -----------------------------------------------------------------------
+    // Helper: push an object into a content array
+    // -----------------------------------------------------------------------
+
+    fn push_obj(parent: &mut Value, obj: Value) {
+        if let Some(arr) = parent["content"].as_array_mut() {
+            arr.push(obj);
+        }
+    }
+
+    // -----------------------------------------------------------------------
+    // id  →  { type:"book", marker:"id", code, content:[desc?] }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_id(&mut self, node: Node, parent: &mut Value) {
+        let mut code: Option<String> = None;
+        let mut desc: Option<String> = None;
+
         let mut cursor = node.walk();
-        if cursor.goto_first_child() {
-            loop {
-                let child = cursor.node();
-                node_2_usj(&child, parent_json_obj, usfm);
-                if !cursor.goto_next_sibling() {
+        for child in node.children(&mut cursor) {
+            match child.kind() {
+                "bookcode"    => code = Some(self.node_text_trimmed(child)),
+                "description" => desc = Some(self.node_text_trimmed(child)),
+                _ => {}
+            }
+        }
+
+        self.parse_state.book_slug = code.clone();
+
+        let mut book_obj = json!({
+            "type":    "book",
+            "marker":  "id",
+            "code":    code.unwrap_or_default(),
+            "content": []
+        });
+
+        if let Some(d) = desc {
+            if !d.is_empty() {
+                Self::push_str(&mut book_obj, d);
+            }
+        }
+
+        Self::push_obj(parent, book_obj);
+    }
+
+    // -----------------------------------------------------------------------
+    // c  →  { type:"chapter", marker:"c", number, sid, altnumber?, pubnumber? }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_c(&mut self, node: Node, parent: &mut Value) {
+        let mut chap_num: Option<String> = None;
+        let mut alt_num:  Option<String> = None;
+        let mut pub_num:  Option<String> = None;
+
+        // Walk one level: c > chapterNumber / ca / cp
+        let mut cursor = node.walk();
+        for child in node.children(&mut cursor) {
+            match child.kind() {
+                "chapterNumber" => chap_num = Some(self.node_text_trimmed(child)),
+                "ca" => {
+                    // ca > chapterNumber
+                    let mut c2 = child.walk();
+                    for sub in child.children(&mut c2) {
+                        if sub.kind() == "chapterNumber" {
+                            alt_num = Some(self.node_text_trimmed(sub));
+                        }
+                    }
+                }
+                "cp" => {
+                    // cp > text
+                    let mut c2 = child.walk();
+                    for sub in child.children(&mut c2) {
+                        if sub.kind() == "text" {
+                            pub_num = Some(self.node_text_trimmed(sub));
+                        }
+                    }
+                }
+                _ => {}
+            }
+        }
+
+        let num = chap_num.unwrap_or_default();
+        let sid = format!(
+            "{} {}",
+            self.parse_state.book_slug.as_deref().unwrap_or(""),
+            num
+        );
+        self.parse_state.current_chapter = Some(num.clone());
+
+        let mut chap_obj = json!({
+            "type":   "chapter",
+            "marker": "c",
+            "number": num,
+            "sid":    sid,
+        });
+
+        if let Some(a) = alt_num.filter(|s| !s.is_empty()) {
+            chap_obj["altnumber"] = Value::String(a);
+        }
+        if let Some(p) = pub_num.filter(|s| !s.is_empty()) {
+            chap_obj["pubnumber"] = Value::String(p);
+        }
+
+        Self::push_obj(parent, chap_obj);
+
+        // cl / cd children are appended to root, not chapter obj
+        let mut cursor2 = node.walk();
+        for child in node.children(&mut cursor2) {
+            if matches!(child.kind(), "cl" | "cd") {
+                self.node_2_usj(child, parent);
+            }
+        }
+    }
+
+    // -----------------------------------------------------------------------
+    // chapter  →  walk children, dispatching c separately
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_chapter(&mut self, node: Node, parent: &mut Value) {
+        let mut cursor = node.walk();
+        for child in node.children(&mut cursor) {
+            if child.kind() == "c" {
+                self.node_2_usj_c(child, parent);
+            } else {
+                self.node_2_usj(child, parent);
+            }
+        }
+    }
+
+    // -----------------------------------------------------------------------
+    // v  →  { type:"verse", marker:"v", number, sid, altnumber?, pubnumber? }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_verse(&mut self, node: Node, parent: &mut Value) {
+        let mut vnum:    Option<String> = None;
+        let mut alt_num: Option<String> = None;
+        let mut pub_num: Option<String> = None;
+
+        let mut cursor = node.walk();
+        for child in node.children(&mut cursor) {
+            match child.kind() {
+                "verseNumber" => {
+                    if vnum.is_none() {
+                        vnum = Some(self.node_text_trimmed(child));
+                    }
+                }
+                "va" => {
+                    let mut c2 = child.walk();
+                    for sub in child.children(&mut c2) {
+                        if sub.kind() == "verseNumber" {
+                            alt_num = Some(self.node_text_trimmed(sub));
+                        }
+                    }
+                }
+                "vp" => {
+                    let mut c2 = child.walk();
+                    for sub in child.children(&mut c2) {
+                        if sub.kind() == "text" {
+                            pub_num = Some(self.node_text_trimmed(sub));
+                        }
+                    }
+                }
+                _ => {}
+            }
+        }
+
+        let num = match vnum {
+            Some(n) => n,
+            None    => return,
+        };
+
+        let sid = format!(
+            "{} {}:{}",
+            self.parse_state.book_slug.as_deref().unwrap_or(""),
+            self.parse_state.current_chapter.as_deref().unwrap_or(""),
+            num
+        );
+
+        let mut v_obj = json!({
+            "type":   "verse",
+            "marker": "v",
+            "number": num,
+            "sid":    sid,
+        });
+
+        if let Some(a) = alt_num.filter(|s| !s.is_empty()) {
+            v_obj["altnumber"] = Value::String(a);
+        }
+        if let Some(p) = pub_num.filter(|s| !s.is_empty()) {
+            v_obj["pubnumber"] = Value::String(p);
+        }
+
+        Self::push_obj(parent, v_obj);
+    }
+
+    // -----------------------------------------------------------------------
+    // ca / va  →  { type:"char", marker, altnumber }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_ca_va(&mut self, node: Node, parent: &mut Value) {
+        let style = node.kind().to_string();
+        let mut alt_num = String::new();
+
+        let mut cursor = node.walk();
+        for child in node.children(&mut cursor) {
+            if matches!(child.kind(), "chapterNumber" | "verseNumber") {
+                alt_num = self.node_text_trimmed(child);
+            }
+        }
+
+        let obj = json!({
+            "type":      "char",
+            "marker":    style,
+            "altnumber": alt_num,
+        });
+        Self::push_obj(parent, obj);
+    }
+
+    // -----------------------------------------------------------------------
+    // paragraph / q / w  →  { type:"para", marker, content:[] }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_para(&mut self, node: Node, parent: &mut Value) {
+        // Handle block wrappers
+        if let Some(first) = node.child(0) {
+            if first.kind().ends_with("Block") {
+                let mut cursor = first.walk();
+                let children: Vec<Node> = first.children(&mut cursor).collect();
+                for child in children {
+                    self.node_2_usj_para(child, parent);
+                }
+                return;
+            }
+        }
+
+        match node.kind() {
+            "paragraph" => {
+                // Find the first non-block child — that is the para marker node
+                let mut cursor = node.walk();
+                let para_marker_node = node
+                    .children(&mut cursor)
+                    .find(|c| !c.kind().ends_with("Block"));
+
+                let para_marker_node = match para_marker_node {
+                    Some(n) => n,
+                    None    => return,
+                };
+                let para_marker = para_marker_node.kind().to_string();
+
+                if para_marker.ends_with("Block") {
+                    return;
+                }
+
+                if para_marker == "b" {
+                    Self::push_obj(parent, json!({ "type": "para", "marker": "b" }));
+                    return;
+                }
+
+                let mut para_obj = json!({
+                    "type":    "para",
+                    "marker":  para_marker,
+                    "content": []
+                });
+
+                let mut c2 = para_marker_node.walk();
+                let children: Vec<Node> = para_marker_node.children(&mut c2).collect();
+                for child in children {
+                    self.node_2_usj(child, &mut para_obj);
+                }
+                Self::push_obj(parent, para_obj);
+            }
+
+            "pi" | "ph" => {
+                let tag_node = match node.child(0) {
+                    Some(n) => n,
+                    None    => return,
+                };
+                let raw = self.node_text(tag_node);
+                let para_marker = raw.replace('\\', "").trim().to_string();
+
+                let mut para_obj = json!({
+                    "type":    "para",
+                    "marker":  para_marker,
+                    "content": []
+                });
+
+                let mut cursor = node.walk();
+                let children: Vec<Node> = node.children(&mut cursor).skip(1).collect();
+                for child in children {
+                    self.node_2_usj(child, &mut para_obj);
+                }
+                Self::push_obj(parent, para_obj);
+            }
+
+            _ => {}
+        }
+    }
+
+    // -----------------------------------------------------------------------
+    // footnotes / crossrefs  →  { type:"note", marker, caller, content:[] }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_notes(&mut self, node: Node, parent: &mut Value) {
+        let tag_node    = match node.child(0) { Some(n) => n, None => return };
+        let caller_node = match node.child(1) { Some(n) => n, None => return };
+
+        let style = self.node_text(tag_node)
+            .replace('\\', "")
+            .trim()
+            .to_string();
+        let caller = self.node_text_trimmed(caller_node);
+
+        let mut note_obj = json!({
+            "type":    "note",
+            "marker":  style,
+            "caller":  caller,
+            "content": []
+        });
+
+        let child_count = node.child_count();
+        for i in 2..child_count.saturating_sub(1) {
+            if let Some(child) = node.child(i.try_into().unwrap()) {
+                self.node_2_usj(child, &mut note_obj);
+            }
+        }
+
+        Self::push_obj(parent, note_obj);
+    }
+
+    // -----------------------------------------------------------------------
+    // char styles  →  { type:"char", marker, content:[] }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_char(&mut self, node: Node, parent: &mut Value) {
+        let tag_node = match node.child(0) { Some(n) => n, None => return };
+
+        // Trim closing tag nodes from the range
+        let mut children_range = node.child_count();
+        for i in (1..node.child_count()).rev() {
+            if let Some(c) = node.child(i.try_into().unwrap()) {
+                let t = c.kind();
+                if t.starts_with('\\') || t == "*" || t.ends_with("Tag") {
+                    children_range -= 1;
+                } else {
                     break;
                 }
             }
         }
-    }
-    else {
-    // Do nothing for all other cases
-    }
-}
-pub fn node_2_usj_id(
-    //verified
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    let query_source = r#"
-            (id
-                (bookcode) @book-code
-                (description)? @desc
-            )
-            "#;
-    //let node.kind() = node.kind();
-    let node_text = node
-        .utf8_text(usfm.as_bytes())
-        .expect("Failed to get node text")
-        .to_string();
-    ////println!("nodeTYPE={}", node_text);
 
-    let query =
-        Query::new(&tree_sitter_usfm3::language(), query_source).expect("Failed to create query");
-    let mut query_cursor = QueryCursor::new();
-    let mut captures = query_cursor.matches(&query, *node, node_text.as_bytes());
-
-    let mut code = None;
-    let mut desc = None;
-
-    while let Some(capture) = captures.next() {
-        // Capture the chapter number
-        if let Some(code_capture) = capture.captures.get(0) {
-            if let Ok(code_text) = code_capture.node.utf8_text(usfm.as_bytes()) {
-                code = Some(code_text.trim().to_string());
-            }
-        }
-
-        // Capture the altTrevorernative chapter number if it exists
-        if let Some(desc_capture) = capture.captures.get(1) {
-            if let Ok(desc_text) = desc_capture.node.utf8_text(usfm.as_bytes()) {
-                desc = Some(desc_text.trim().to_string());
-            }
-        }
-    }
-
-    let mut book_json_obj = json!({
-        "type": "book",
-        "marker": "id",
-        "code": code.clone().unwrap_or_default(),
-        "content": if let Some(d) = desc {
-            if d.is_empty() {
-                Vec::<String>::new() // Return an empty Vec if desc is an empty string
-            } else {
-                vec![d] // Wrap desc in a Vec if it's not empty
-            }
-        } else {
-            Vec::<String>::new() // Return an empty Vec if desc is None
-        },
-    });
-
-    parent_json_obj.push(book_json_obj.clone());
-}
-
-
-pub fn node_2_usj_c(
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    let query_source = r#"
-        (c
-            (chapterNumber) @chap-num
-            (ca (chapterNumber) @alt-num)?
-            (cp (text) @pub-num)?
-        )
-    "#;
-    let query =
-        Query::new(&tree_sitter_usfm3::language(), query_source).expect("Failed to create query");
-    let mut query_cursor = QueryCursor::new();
-
-    let mut captures = query_cursor.matches(&query, *node, usfm.as_bytes());
-
-    let mut chapter_number: Option<String> = None;
-    let mut alt_number: Option<String> = None;
-    let mut publication_number: Option<String> = None;
-    let mut chap_ref: Option<String> = None;
-    let mut global_chapter = CHAPTER_NUMBER.lock().unwrap();
-    while let Some(capture) = captures.next() {
-        if let Some(chap_num_capture) = capture.captures.get(0) {
-            if let Ok(num) = chap_num_capture.node.utf8_text(usfm.as_bytes()) {
-                chapter_number = Some(num.trim().to_string());
-            }
-        }
-
-        if let Some(alt_num_capture) = capture.captures.get(1) {
-            if let Ok(alt) = alt_num_capture.node.utf8_text(usfm.as_bytes()) {
-                alt_number = Some(alt.trim().to_string());
-            }
-        }
-
-        if let Some(pub_num_capture) = capture.captures.get(2) {
-            if let Ok(pub_num) = pub_num_capture.node.utf8_text(usfm.as_bytes()) {
-                publication_number = Some(pub_num.trim().to_string());
-            }
-        }
-    }
-
-    for child in &*parent_json_obj {
-        if child["type"] == "book" {
-            if let Some(code) = child.get("code") {
-                chap_ref = Some(format!(
-                    "{} {}",
-                    code.as_str().unwrap_or(""),
-                    chapter_number.clone().unwrap_or_default()
-                ));
-            }
-            break;
-        }
-    }
-    *global_chapter = chap_ref.clone();
-
-    let mut chap_json_obj = json!({
-        "type": "chapter",
-        "marker": "c",
-        "number": chapter_number.clone().unwrap_or_default(),
-        "sid": chap_ref.clone().unwrap_or_default(),
-    });
-
-    if let Some(alt) = alt_number {
-        chap_json_obj["altnumber"] = json!(alt);
-    }
-    if let Some(pub_num) = publication_number {
-        chap_json_obj["pubnumber"] = json!(pub_num);
-    }
-
-    parent_json_obj.push(chap_json_obj);
-
-    for child in node.children(&mut node.walk()) {
-        if child.kind() == "c" {
-            node_2_usj_ca_va(&child, parent_json_obj, usfm);
-        } else {
-            node_2_usj(&child, parent_json_obj, usfm);
-        }
-    }
-}
-fn node_2_usj_chapter(
-        node: &tree_sitter::Node,
-        parent_json_obj: &mut Vec<serde_json::Value>,
-        usfm: &str,
-) {
-    for child in node.children(&mut node.walk()) {
-        if child.kind() == "c" {
-            node_2_usj_c(&child, parent_json_obj, usfm);
-        } else {
-            node_2_usj(&child, parent_json_obj, usfm);
-        }
-    }
-}
-pub fn node_2_usj_ca_va(
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    // Get the first child node to determine the style
-    let tag_node = node
-        .named_child(0)
-        .expect("Expected a child node for style");
-    let style = tag_node.kind();
-    //println!("STYLE IS........{}", style);
-    // Clean up the style string
-    let style = if style.starts_with('\\') {
-        style.replace('\\', "").trim().to_string()
-    } else {
-        node.kind().to_string()
-    };
-
-    let mut children_range_start = 1;
-
-    // Check if the second child is a numbered style
-    if node.named_child_count() > 1
-        && node
-            .named_child(1)
-            .expect("Expected a numbered child")
-            .kind()
-            .starts_with("numbered")
-    {
-        let num_node = node.named_child(1).expect("Expected a numbered child");
-        let num = num_node
-            .utf8_text(usfm.as_bytes())
-            .expect("Failed to get number text")
+        let style = self.node_text(tag_node)
+            .replace('\\', "")
+            .replace('+', "")
+            .trim()
             .to_string();
-        let style = format!("{}{}", style, num);
-        children_range_start = 2; // Start from the third child
-    }
 
-    // Create the paragraph JSON object
-    let mut para_json_obj = json!({
-        "type": "char",
-        "marker": style,
-        "content": [],
-    });
-
-    // Append the paragraph object to the parent parent_json_obj
-    parent_json_obj.push(para_json_obj.clone());
-
-    // Create a TreeCursor to iterate through the children
-    let mut cursor = node.walk();
-    cursor.goto_first_child(); // Move to the first child
-
-    // Skip to the starting index for children
-    for _ in 0..children_range_start {
-        cursor.goto_next_sibling(); // Move to the next sibling
-    }
-
-    // Process the remaining children
-    while cursor.goto_next_sibling() {
-        let child = cursor.node();
-        let child_type = child.kind();
-        if child_type == "text"
-            || child_type == "footnote"
-            || child_type == "crossref"
-            || child_type == "verseText"
-            || child_type == "v"
-            || child_type == "b"
-            || child_type == "milestone"
-            || child_type == "zNameSpace"
-        {
-            // Only nest these types inside the upper para style node
-            node_2_usj(
-                &child,
-                para_json_obj["content"].as_array_mut().unwrap(),
-                usfm,
-            );
-        } else {
-            node_2_usj(&child, parent_json_obj, usfm);
-        }
-    }
-}
-pub fn node_2_usj_verse(
-    //verified
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-    chapter_number: &Option<String>,
-) {
-    // Create a query to capture verse information
-
-    //    print_node(node, usfm, 0);
-    let query_source = r#"
-    (v
-        (verseNumber) @vnum
-        (va (verseNumber) @alt)?
-        (vp (text) @vp)?
-    )
-    "#;
-
-    ////println!("{:?}", chapter);
-    let query =
-        Query::new(&tree_sitter_usfm3::language(), query_source).expect("Failed to create query");
-    let mut query_cursor = QueryCursor::new();
-
-    // Execute the query against the current node
-    let mut captures = query_cursor.matches(&query, *node, usfm.as_bytes());
-
-    let mut verse_number = None;
-    let mut alt_number = None;
-    let mut publication_number = None;
-    let mut verse_text = String::new();
-
-    // Iterate over the captures returned by the query
-    while let Some(capture) = captures.next() {
-        // Capture the verse number
-        if let Some(vnum_capture) = capture.captures.get(0) {
-            if let Ok(num) = vnum_capture.node.utf8_text(usfm.as_bytes()) {
-                verse_number = Some(num.trim().to_string());
-            }
-        }
-
-        // Capture the alternative verse number if it exists
-        if let Some(alt_capture) = capture.captures.get(1) {
-            if let Ok(alt) = alt_capture.node.utf8_text(usfm.as_bytes()) {
-                alt_number = Some(alt.trim().to_string());
-            }
-        }
-
-        // Capture the publication number if it exists
-        if let Some(pub_capture) = capture.captures.get(2) {
-            if let Ok(pub_num) = pub_capture.node.utf8_text(usfm.as_bytes()) {
-                publication_number = Some(pub_num.trim().to_string());
-            }
-        }
-    }
-
-    // Create the verse JSON object
-    let ref_sid = format!(
-        "{}:{}",
-        chapter_number.as_ref().unwrap_or(&"0".to_string()),
-        verse_number.clone().unwrap_or_default()
-    );
-    let mut verse_json_obj = json!({
-        "type": "verse",
-        "marker": "v",
-        "number": verse_number.clone().unwrap_or_default(),
-        "sid": ref_sid
-    });
-
-    // Add alternative and publication numbers if they exist
-    if let Some(alt) = alt_number {
-        verse_json_obj["altnumber"] = json!(alt);
-    }
-    if let Some(pub_num) = publication_number {
-        verse_json_obj["pubnumber"] = json!(pub_num);
-    }
-
-    parent_json_obj.push(verse_json_obj);
-}
-fn node_2_usj_para(
-    //outdated version
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    // if node.child_count() == 0 {
-    //     return;
-    // }
-    // let mut cursor = node.walk();
-    // cursor.goto_first_child(); // Move to the first child
-
-    // while cursor.goto_next_sibling() {
-    //     let current_node = cursor.node();
-    //     return;
-    // }
-    // Create a TreeCursor to iterate through the children
-    let mut cursor = node.walk();
-    //cursor.goto_first_child(); // Move to the first child
-
-    // Check if the first child is a block type
-    if cursor.node().kind().ends_with("Block") {
-        // cursor.goto_first_child(); // Move to the first child of the block
-        cursor.goto_next_sibling();
-        let child = cursor.node();
-        node_2_usj(&child, parent_json_obj, usfm); // Recursively process child nodes
-    } else if node.kind() == "paragraph" {
-        // Extract the paragraph marker using a query
-        let query_source = r#"(paragraph (_) @para-marker)"#;
-        let query = Query::new(&tree_sitter_usfm3::language(), query_source)
-            .expect("Failed to create query");
-        let mut query_cursor = QueryCursor::new();
-        let mut captures = query_cursor.matches(&query, *node, usfm.as_bytes());
-
-        // Process the captures to get the paragraph marker
-        if let Some(capture) = captures.next() {
-            let para_marker = capture.captures[0].node.kind(); // Get the marker type
-            let mut para_json_obj = json!({
-                "type": "para",
-                "marker": para_marker,
-                "content": []
-            });
-
-            // Create a TreeCursor to iterate through the children of the capture
-            let mut child_cursor = capture.captures[0].node.walk();
-            child_cursor.goto_first_child(); // Move to the first child
-
-            // Process the children of the paragraph
-            while child_cursor.goto_next_sibling() {
-                let child = child_cursor.node();
-                node_2_usj(
-                    &child,
-                    para_json_obj["content"].as_array_mut().unwrap(),
-                    usfm,
-                );
-            }
-
-            // Append the paragraph JSON object to the parent parent_json_obj
-            parent_json_obj.push(para_json_obj);
-        }
-    } else if node.kind() == "pi" || node.kind() == "ph" {
-        // Extract the marker for pi or ph
-        let para_marker = node
-            .utf8_text(usfm.as_bytes())
-            .expect("Failed to get node text")
-            .trim_start_matches('\\')
-            .to_string();
-        let mut para_json_obj = json!({
-            "type": "para",
-            "marker": para_marker,
-            "content": [],
+        let mut char_obj = json!({
+            "type":    "char",
+            "marker":  style,
+            "content": []
         });
 
-        // Create a TreeCursor to iterate through the children
-        let mut child_cursor: tree_sitter::TreeCursor<'_> = node.walk();
-        //child_cursor.goto_first_child(); // Move to the first child
-
-        // Process the remaining children
-        while child_cursor.goto_next_sibling() {
-            let child = child_cursor.node();
-            node_2_usj(
-                &child,
-                para_json_obj["parent_json_obj"].as_array_mut().unwrap(),
-                usfm,
-            );
-        }
-
-        // Append the paragraph JSON object to the parent parent_json_obj
-        parent_json_obj.push(para_json_obj);
-    }
-}
-
-pub fn node_2_usj_notes(
-    //verified
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    // Collect the children into a vector
-    let children: Vec<_> = node.children(&mut node.walk()).collect();
-
-    // Get the tag node and caller node
-    let tag_node = &children[0];
-    let caller_node = &children[1];
-
-    // Extract the style from the tag node
-    let style = usfm[tag_node.start_byte()..tag_node.end_byte()]
-        .replace("\\", "")
-        .trim()
-        .to_string();
-
-    // Create the note JSON object
-    let mut note_json_obj = json!({
-        "type": "note",
-        "marker": style,
-        "content": [],
-    });
-
-    // Extract the caller text
-    let caller_text = usfm[caller_node.start_byte()..caller_node.end_byte()]
-        .trim()
-        .to_string();
-    note_json_obj["caller"] = json!(caller_text);
-
-    // Process the remaining children (from index 2 to the last child)
-    for child in &children[2..] {
-        node_2_usj(
-            child,
-            note_json_obj["content"].as_array_mut().unwrap(),
-            usfm,
-        );
-    }
-
-    // Append the note JSON object to the parent parent_json_obj
-    parent_json_obj.push(note_json_obj);
-}
-pub fn node_2_usj_char(
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>, // Change back to Vec<Value>
-    usfm: &str,
-) {
-    // Ensure the node has children
-    if node.child_count() < 1 {
-        return;
-    }
-
-    // Get the tag node (first child)
-    let tag_node = node.children(&mut node.walk()).next().unwrap();
-    //println!("tag node.............{}",tag_node);
-    // Determine the range of children to process
-    let children: Vec<_> = node.children(&mut node.walk()).collect();
-    let mut children_range = children.len();
-    //println!("count child node.............{}", children_range);
-    if let Some(last_child) = node.child(children_range - 1) {
-        if last_child.kind().starts_with('\\') {
-            children_range -= 1; // Exclude the closing node if it starts with '\'
-        }
-    }
-
-    // Extract the style from the tag node
-    let style = tag_node
-        .utf8_text(usfm.as_bytes())
-        .expect("Failed to get tag node text")
-        .replace("\\", "")
-        .replace("+", "")
-        .trim()
-        .to_string();
-    // let node_text = node
-    // .utf8_text(usfm.as_bytes())
-    // .expect("Failed to get node text")
-    // .replace('\n', "")
-    // .to_string();
-
-    // // Create a regex to match any prefix starting with a backslash followed by any characters
-    // let re = Regex::new(r"\\[^ |]*").unwrap(); // Matches a backslash followed by any characters until a space
-
-    // Remove the unwanted prefix from the node_text
-    // let cleaned_text = re.replace_all(&node_text, "").trim().to_string();
-    // let cleaned_text = cleaned_text.split('|').next().unwrap_or(&cleaned_text).trim().to_string();
-
-    // let marker_parts: Vec<&str> = style.split_whitespace().collect();
-    // let marker = marker_parts.get(0).unwrap_or(&"").to_string(); // Get the first part as the marker
-    // let content_value = if cleaned_text.is_empty() {
-    //  Vec::<String>::new() // Return an empty Vec if cleaned_text is an empty string
-    // } else {
-    //     vec![cleaned_text] // Wrap cleaned_text in a Vec if it's not empty
-    // };
-    // Create the character JSON object
-    let mut char_json_obj = json!({
-        "type": "char",
-        "marker": style,
-        "content": [],
-    });
-
-    // Get a mutable reference to the content array
-    let content_array = char_json_obj["content"]
-        .as_array_mut()
-        .expect("Expected content to be an array");
-
-    // Process child nodes (excluding the first and possibly the last)
-    // for i in 0..children_range {
-    //     if let Some(child) = node.child(i) {
-    //         node_2_usj(&child, content_array, usfm); // Pass the mutable reference to the content array
-    //     }
-    // }
-    // let mut child_cursor = node.walk();
-    // child_cursor.goto_first_child(); // Move to the first child
-
-    // // Process the remaining children
-    for child in &children[1..children_range] {
-        node_2_usj(child, content_array, usfm);
-    }
-    // while child_cursor.goto_next_sibling() {
-    //     let child = child_cursor.node();
-    //     node_2_usj(
-    //         &child,
-    //         &mut char_json_obj["content"].as_array_mut().unwrap(),
-    //         usfm,
-    //     );
-    // }
-
-    // Append the character JSON object to the parent JSON object
-    parent_json_obj.push(char_json_obj); // Push the character JSON object directly to the Vec
-
-    let mut cursor = node.walk();
-    cursor.goto_first_child(); // Move to the first child
-
-    while cursor.goto_next_sibling() {
-        let current_node = cursor.node();
-        return;
-    }
-}
-
-pub fn node_2_usj_attrib(
-    //verified
-    node: &tree_sitter::Node,
-    content: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    // Get the attribute name node
-    let attrib_name_node = node.child(0).expect("Node should have at least one child");
-    let mut attrib_name = attrib_name_node
-        .utf8_text(usfm.as_bytes())
-        .unwrap()
-        .trim()
-        .to_string();
-
-    // Handle special cases for attribute names
-    if attrib_name == "|" {
-        attrib_name = DEFAULT_ATTRIB_MAP
-            .iter()
-            .find(|&&(key, _)| key == node.parent().unwrap().kind())
-            .map(|&(_, value)| value.to_string())
-            .unwrap_or(attrib_name);
-    }
-    if attrib_name == "src" {
-        attrib_name = "file".to_string();
-    }
-
-    // Query to capture attribute values
-    let query_source = r#"
-    ((attributeValue) @attrib-val)
-    "#;
-
-    let query =
-        Query::new(&tree_sitter_usfm3::language(), query_source).expect("Failed to create query");
-    let mut cursor = QueryCursor::new();
-
-    // Execute the query against the current node
-    let mut captures = cursor.matches(&query, *node, usfm.as_bytes());
-
-    // Initialize attrib_value
-    let attrib_value: String;
-
-    // Capture the attribute value
-    let mut value = "";
-    if let Some(capture) = captures.next() {
-        if let Some(attrib_value_capture) = capture.captures.get(0) {
-            value = attrib_value_capture
-                .node
-                .utf8_text(usfm.as_bytes())
-                .unwrap()
-                .trim();
-        }
-    }
-    if value.len() > 0 {
-        attrib_value = value.to_string(); // Assign the string value
-    } else {
-        attrib_value = "".to_string(); // Reset if the number is not greater than 0
-    }
-
-    // Create the JSON object
-    let mut attribute_json_obj = json!({
-        attrib_name: attrib_value
-    });
-   // println!("ATTRIB ::::{}",attrib_value);
-    // Append the JSON object to the content
-    content.push(attribute_json_obj);
-}
-pub fn node_2_usj_table(
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    //let node.kind() = node.kind();
-    match node.kind() {
-        "table" => {
-            // Create a JSON object for the table
-            let mut table_json_obj = json!({
-                "type": "table",
-                "content": [],
-            });
-
-            // Process all children of the table
-            for child in node.named_children(&mut node.walk()) {
-                node_2_usj(
-                    &child,
-                    table_json_obj["content"].as_array_mut().unwrap(),
-                    usfm,
-                );
+        for i in 1..children_range {
+            if let Some(child) = node.child(i.try_into().unwrap()) {
+                self.node_2_usj(child, &mut char_obj);
             }
-
-            // Append the table JSON object to the parent parent_json_obj
-            parent_json_obj.push(table_json_obj);
         }
 
-        "tr" => {
-            // Create a JSON object for the table row
-            let mut row_json_obj = json!({
-                "type": "table:row",
-                "marker": "tr",
-                "content": [],
-            });
+        Self::push_obj(parent, char_obj);
+    }
 
-            // Process all children of the row, skipping the first child
-            for child in node.named_children(&mut node.walk()).skip(1) {
-                node_2_usj(
-                    &child,
-                    row_json_obj["content"].as_array_mut().unwrap(),
-                    usfm,
-                );
+    // -----------------------------------------------------------------------
+    // table / tr / table cells
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_table(&mut self, node: Node, parent: &mut Value) {
+        match node.kind() {
+            "table" => {
+                let mut table_obj = json!({ "type": "table", "content": [] });
+                let mut cursor = node.walk();
+                let children: Vec<Node> = node.children(&mut cursor).collect();
+                for child in children {
+                    self.node_2_usj(child, &mut table_obj);
+                }
+                Self::push_obj(parent, table_obj);
             }
+            "tr" => {
+                let mut row_obj = json!({
+                    "type":    "table:row",
+                    "marker":  "tr",
+                    "content": []
+                });
+                let mut cursor = node.walk();
+                let children: Vec<Node> = node.children(&mut cursor).skip(1).collect();
+                for child in children {
+                    self.node_2_usj(child, &mut row_obj);
+                }
+                Self::push_obj(parent, row_obj);
+            }
+            cell_type if in_set(cell_type, TABLE_CELL_MARKERS) => {
+                let tag_node = match node.child(0) { Some(n) => n, None => return };
+                let style = self.node_text(tag_node)
+                    .replace('\\', "")
+                    .trim()
+                    .to_string();
 
-            // Append the row JSON object to the parent parent_json_obj
-            parent_json_obj.push(row_json_obj);
+                let align = if style.contains("tcc") || style.contains("thc") {
+                    "center"
+                } else if style.contains('r') {
+                    "end"
+                } else {
+                    "start"
+                };
+
+                let mut cell_obj = json!({
+                    "type":    "table:cell",
+                    "marker":  style,
+                    "align":   align,
+                    "content": []
+                });
+
+                let mut cursor = node.walk();
+                let children: Vec<Node> = node.children(&mut cursor).skip(1).collect();
+                for child in children {
+                    self.node_2_usj(child, &mut cell_obj);
+                }
+                Self::push_obj(parent, cell_obj);
+            }
+            _ => {}
         }
+    }
 
-        _ if TABLE_CELL_MARKERS.contains(&node.kind()) => {
-            // Handle table cell markers
-            let tag_node = node.named_child(0).expect("Expected a child node for tag");
-            let style = tag_node
-                .utf8_text(usfm.as_bytes())
-                .expect("Failed to get cell style")
-                .replace("\\", "")
-                .trim()
+    // -----------------------------------------------------------------------
+    // attribute nodes  →  sets a key on the parent object
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_attrib(&mut self, node: Node, parent: &mut Value) {
+        let attrib_name_node = match node.child(0) { Some(n) => n, None => return };
+        let mut attrib_name = self.node_text_trimmed(attrib_name_node);
+
+        if attrib_name == "|" {
+            let parent_type = node.parent().map(|p| p.kind().to_string()).unwrap_or_default();
+            let resolved = parent_type.replace("Nested", "");
+            attrib_name = default_attrib_map(&resolved)
+                .unwrap_or("|")
                 .to_string();
-
-            // Create a JSON object for the table cell
-            let mut cell_json_obj = json!({
-                "type": "table:cell",
-                "marker": style,
-                "content": [],
-            });
-
-            // Set alignment based on the style
-            if style.contains("r") {
-                cell_json_obj["align"] = json!("end");
-            } else {
-                cell_json_obj["align"] = json!("start");
-            }
-
-            // Process all children of the cell, skipping the first child
-            for child in node.named_children(&mut node.walk()).skip(1) {
-                node_2_usj(
-                    &child,
-                    cell_json_obj["content"].as_array_mut().unwrap(),
-                    usfm,
-                );
-            }
-
-            // Append the cell JSON object to the parent parent_json_obj
-            parent_json_obj.push(cell_json_obj);
         }
 
-        _ => {}
+        if attrib_name == "src" {
+            attrib_name = "file".to_string();
+        }
+
+        // Find attributeValue child
+        let mut attrib_value = String::new();
+        let mut cursor = node.walk();
+        for child in node.children(&mut cursor) {
+            if child.kind() == "attributeValue" {
+                attrib_value = self.node_text_trimmed(child);
+                break;
+            }
+        }
+
+        parent[attrib_name] = Value::String(attrib_value);
     }
-}
 
-fn node_2_usj_milestone(
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    // Create a query to capture the milestone or zNameSpace name
-    let query_source = r#"
-[
-    (milestoneTag) @ms-name
-    (milestoneStartTag) @ms-name
-    (milestoneEndTag) @ms-name
-    (zSpaceTag) @ms-name
-]
-"#;
-    let query =
-        Query::new(&tree_sitter_usfm3::language(), query_source).expect("Failed to create query");
-    let mut query_cursor = QueryCursor::new();
+    // -----------------------------------------------------------------------
+    // milestone / zNameSpace  →  { type:"ms", marker, content:[]? }
+    // -----------------------------------------------------------------------
 
-    // Execute the query against the current node
-    let mut captures = query_cursor.matches(&query, *node, usfm.as_bytes());
+    fn node_2_usj_milestone(&mut self, node: Node, parent: &mut Value) {
+        // Find the milestone tag name child
+        let mut style = String::new();
+        let mut cursor = node.walk();
+        for child in node.children(&mut cursor) {
+            let k = child.kind();
+            if matches!(k, "milestoneTag" | "milestoneStartTag" | "milestoneEndTag" | "zSpaceTag") {
+                style = self.node_text(child)
+                    .replace('\\', "")
+                    .trim()
+                    .to_string();
+                break;
+            }
+        }
+        if style.is_empty() { return; }
 
-    // Extract the style from the captures
-    let style = if let Some(capture) = captures.next() {
-        if let Some(ms_name_capture) = capture.captures.get(0) {
-            let style = ms_name_capture
-                .node
-                .utf8_text(usfm.as_bytes())
-                .expect("Failed to get milestone or zNameSpace text")
-                .replace("\\", "")
-                .trim()
-                .to_string();
-            style
+        let mut ms_obj = json!({
+            "type":    "ms",
+            "marker":  style,
+            "content": []
+        });
+
+        let mut cursor2 = node.walk();
+        for child in node.children(&mut cursor2) {
+            if child.kind().ends_with("Attribute") {
+                self.node_2_usj(child, &mut ms_obj);
+            }
+        }
+
+        // Remove content if empty
+        if ms_obj["content"].as_array().map(|a| a.is_empty()).unwrap_or(false) {
+            ms_obj.as_object_mut().unwrap().remove("content");
+        }
+
+        Self::push_obj(parent, ms_obj);
+    }
+
+    // -----------------------------------------------------------------------
+    // esb / cat / fig / ref
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_special(&mut self, node: Node, parent: &mut Value) {
+        match node.kind() {
+            "esb" => {
+                let mut sidebar_obj = json!({
+                    "type":    "sidebar",
+                    "marker":  "esb",
+                    "content": []
+                });
+                let count = node.child_count();
+                for i in 1..count.saturating_sub(1) {
+                    if let Some(child) = node.child(i.try_into().unwrap()) {
+                        self.node_2_usj(child, &mut sidebar_obj);
+                    }
+                }
+                Self::push_obj(parent, sidebar_obj);
+            }
+            "cat" => {
+                // Find category child
+                let mut cursor = node.walk();
+                for child in node.children(&mut cursor) {
+                    if child.kind() == "category" {
+                        let category = self.node_text_trimmed(child);
+                        parent["category"] = Value::String(category);
+                        break;
+                    }
+                }
+            }
+            "fig" => {
+                let mut fig_obj = json!({
+                    "type":    "figure",
+                    "marker":  "fig",
+                    "content": []
+                });
+                let count = node.child_count();
+                for i in 1..count.saturating_sub(1) {
+                    if let Some(child) = node.child(i.try_into().unwrap()) {
+                        self.node_2_usj(child, &mut fig_obj);
+                    }
+                }
+                Self::push_obj(parent, fig_obj);
+            }
+            "ref" => {
+                let mut ref_obj = json!({ "type": "ref", "content": [] });
+                let count = node.child_count();
+                for i in 1..count.saturating_sub(1) {
+                    if let Some(child) = node.child(i.try_into().unwrap()) {
+                        self.node_2_usj(child, &mut ref_obj);
+                    }
+                }
+                Self::push_obj(parent, ref_obj);
+            }
+            _ => {}
+        }
+    }
+
+    // -----------------------------------------------------------------------
+    // Generic para-style markers  →  { type:"para", marker, content:[] }
+    // -----------------------------------------------------------------------
+
+    fn node_2_usj_generic(&mut self, node: Node, parent: &mut Value) {
+        let tag_node = match node.child(0) { Some(n) => n, None => return };
+
+        let raw = self.node_text(tag_node);
+        let mut style = if raw.starts_with('\\') {
+            raw.replace('\\', "").trim().to_string()
         } else {
-            String::new()
-        }
-    } else {
-        String::new()
-    };
+            node.kind().to_string()
+        };
 
-    // Create the milestone or zNameSpace JSON object
-    let mut ms_json_obj = json!({
-        "type": "ms",
-        "marker": style,
-        "content": [],
-    });
-
-    // Create a TreeCursor to iterate through the children
-    let mut cursor = node.walk();
-    cursor.goto_first_child(); // Move to the first child
-
-    // Process all children
-    while cursor.goto_next_sibling() {
-        let child = cursor.node();
-        if child.kind().ends_with("Attribute") {
-            node_2_usj(&child, ms_json_obj["content"].as_array_mut().unwrap(), usfm);
-        }
-    }
-
-    // If the content is empty, remove the content field
-    if ms_json_obj["content"].as_array().unwrap().is_empty() {
-        ms_json_obj.as_object_mut().unwrap().remove("content");
-    }
-
-    // Append the milestone or zNameSpace JSON object to the parent content
-    parent_json_obj.push(ms_json_obj);
-}
-
-pub fn node_2_usj_special(
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    // let node.kind() = node.kind();
-    match node.kind() {
-        "esb" => {
-            // Create a JSON object for the sidebar
-            let mut sidebar_json_obj = json!({
-                "type": "sidebar",
-                "marker": "esb",
-                "content": [],
-            });
-
-            // Create a TreeCursor to iterate through the children
-            let mut cursor = node.walk();
-            cursor.goto_first_child(); // Move to the first child
-
-            let mut index = 0; // Initialize an index counter
-
-            // Process all children except the first and last
-            while cursor.goto_next_sibling() {
-                let child = cursor.node();
-                index += 1; // Increment the index counter
-
-                // Skip the first and last child
-                if index > 0 && index < node.named_child_count() - 1 {
-                    node_2_usj(
-                        &child,
-                        sidebar_json_obj["content"].as_array_mut().unwrap(),
-                        usfm,
-                    );
-                }
-            }
-
-            // Append the sidebar JSON object to the parent parent_json_obj
-            parent_json_obj.push(sidebar_json_obj);
-        }
-
-        "cat" => {
-            // Create a query to capture the category
-            let query_source = r#"
-    (category) @category
-    "#;
-            let query = Query::new(&tree_sitter_usfm3::language(), query_source)
-                .expect("Failed to create query");
-            let mut query_cursor = QueryCursor::new();
-
-            // Execute the query against the current node
-            let mut captures = query_cursor.matches(&query, *node, usfm.as_bytes());
-
-            // Extract the category from the captures
-            if let Some(capture) = captures.next() {
-                if let Some(category_capture) = capture.captures.get(0) {
-                    let category = category_capture
-                        .node
-                        .utf8_text(usfm.as_bytes())
-                        .expect("Failed to get category text")
-                        .trim()
-                        .to_string();
-                    // Add the category to the parent JSON object
-                    parent_json_obj.push(json!({
-                        "type": "category",
-                        "marker": "cat",
-                        "category": category,
-                    }));
-                }
+        let mut children_range_start = 1usize;
+        if let Some(second) = node.child(1) {
+            if second.kind().starts_with("numbered") {
+                style.push_str(&self.node_text(second));
+                children_range_start = 2;
             }
         }
 
-        "fig" => {
-            // Create a JSON object for the figure
-            let mut fig_json_obj = json!({
-                "type": "figure",
-                "marker": "fig",
-                "content": [],
-            });
+        let mut para_obj = json!({
+            "type":    "para",
+            "marker":  style.trim(),
+            "content": []
+        });
+        let mut node_added = false;
 
-            // Create a TreeCursor to iterate through the children
-            let mut cursor = node.walk();
-            cursor.goto_first_child(); // Move to the first child
+        let count = node.child_count();
+        for i in children_range_start..count {
+            if let Some(child) = node.child(i.try_into().unwrap()) {
+                let ct = child.kind();
+                let nestable = in_set(ct, CHAR_STYLE_MARKERS)
+                    || is_nested_char(ct)
+                    || in_set(ct, OTHER_PARA_NESTABLES);
 
-            let mut index = 0; // Initialize an index counter
-
-            // Process all children except the first and last
-            while cursor.goto_next_sibling() {
-                let child = cursor.node();
-                index += 1; // Increment the index counter
-
-                // Skip the first and last child
-                if index > 0 && index < node.named_child_count() - 1 {
-                    node_2_usj(
-                        &child,
-                        fig_json_obj["content"].as_array_mut().unwrap(),
-                        usfm,
-                    );
+                if nestable {
+                    self.node_2_usj(child, &mut para_obj);
+                } else {
+                    if !node_added {
+                        Self::push_obj(parent, para_obj.clone());
+                        node_added = true;
+                    } 
+                    self.node_2_usj(child, parent);
                 }
             }
-
-            // Append the figure JSON object to the parent parent_json_obj
-            parent_json_obj.push(fig_json_obj);
         }
-
-        _ => {}
-    }
-}
-pub fn node_2_usj_generic(
-    node: &tree_sitter::Node,
-    parent_json_obj: &mut Vec<serde_json::Value>,
-    usfm: &str,
-) {
-    // Get the first child node to determine the style
-    //let tag_node = node.child(0).expect("Expected a tag node");
-    // let mut binding=node.walk();
-    // let tag_node=node.children(&mut binding).next().unwrap();
-    let mut style_cap = node
-        .utf8_text(usfm.as_bytes())
-        .expect("Failed to get tag node text")
-        .replace("\\", "")
-        .replace("+", "")
-        .trim()
-        .to_string();
-    //println!("style_cap1:{:?}",style_cap);
-    // Clean up the style string
-
-    let style = style_cap.replace('\\', "").trim().to_string();
-    // } else {
-    //     node.kind().to_string()
-    // };
-    // println!("style:{}",style_cap);
-    let mut children_range_start = 1;
-
-    // Check if the second child is a numbered style
-    if node.named_child_count() > 1
-        && node
-            .named_child(1)
-            .expect("Expected a numbered child")
-            .kind()
-            .starts_with("numbered")
-    {
-        let num_node = node.named_child(1).expect("Expected a numbered child");
-        let num = num_node
-            .utf8_text(usfm.as_bytes())
-            .expect("Failed to get number text")
-            .to_string();
-        let style = format!("{}{}", style, num);
-        children_range_start = 2; // Start from the third child
-    }
-    // let mut text = node.text()
-    //     .utf8_text(usfm.as_bytes())
-    //     .expect("Failed to get node text")
-    //     .replace('\n', "")
-    //     .to_string();
-    // Create the paragraph JSON object
-    let node_text = node
-        .utf8_text(usfm.as_bytes())
-        .expect("Failed to get node text")
-        .replace('\n', "")
-        .to_string();
-
-    // Create a regex to match any prefix starting with a backslash followed by any characters
-    let re = Regex::new(r"\\[^ ]*").unwrap(); // Matches a backslash followed by any characters until a space
-
-    // Remove the unwanted prefix from the node_text
-    let cleaned_text = re.replace_all(&node_text, "").trim().to_string();
-
-    let marker_parts: Vec<&str> = style.split_whitespace().collect();
-    let marker = marker_parts.get(0).unwrap_or(&"").to_string(); // Get the first part as the marker
-    let content_value = if cleaned_text.is_empty() {
-        Vec::<String>::new() // Return an empty Vec if cleaned_text is an empty string
-    } else {
-        vec![cleaned_text] // Wrap cleaned_text in a Vec if it's not empty
-    };
-
-    // Create the JSON object
-    let mut para_json_obj = json!({
-        "content": content_value,
-        "marker": marker,
-        "type": "para",
-    });
-
-    // Append the paragraph object to the parent parent_json_obj
-    parent_json_obj.push(para_json_obj.clone());
-    let mut cursor = node.walk();
-    cursor.goto_first_child(); // Move to the first child
-
-    while cursor.goto_next_sibling() {
-        let current_node = cursor.node();
-        return;
-    }
-    // Process the remaining children
-    for child in node.children(&mut node.walk()) {
-        // Check if the child's kind is in the specified markers
-        if CHAR_STYLE_MARKERS.contains(&child.kind())
-            || NESTED_CHAR_STYLE_MARKERS.contains(&child.kind())
-            || [
-                "text",
-                "footnote",
-                "crossref",
-                "verseText",
-                "v",
-                "b",
-                "milestone",
-                "zNameSpace",
-            ]
-            .contains(&child.kind())
-        {
-            println!("child kind: {}", child.kind());
-            // Only nest these types inside the upper para style node
-            node_2_usj(
-                &child,
-                para_json_obj["content"].as_array_mut().unwrap(),
-                usfm,
-            );
-        } else {
-            node_2_usj(&child, parent_json_obj, usfm);
+        if !node_added {
+            Self::push_obj(parent, para_obj);
         }
     }
-}
 
-fn print_node(node: &tree_sitter::Node, usfm: &str, depth: usize) {
-    let indent = "  ".repeat(depth);
-    let node_text = node.utf8_text(usfm.as_bytes()).unwrap_or_default();
-    println!("{}Node: {}, Text: {}", indent, node.kind(), node_text);
+    // -----------------------------------------------------------------------
+    // text  →  plain string appended to content
+    // -----------------------------------------------------------------------
 
-    for child in node.children(&mut node.walk()) {
-        print_node(&child, usfm, depth + 1);
+    fn push_text_node(&mut self, node: Node, parent: &mut Value) {
+        let text = self.node_text(node).replace('~', " ");
+        if !text.is_empty() {
+            Self::push_str(parent, text);
+        }
+    }
+
+    // -----------------------------------------------------------------------
+    // verseText  →  recurse into children
+    // -----------------------------------------------------------------------
+
+    fn handle_verse_text(&mut self, node: Node, parent: &mut Value) {
+        let mut cursor = node.walk();
+        let children: Vec<Node> = node.children(&mut cursor).collect();
+        for child in children {
+            self.node_2_usj(child, parent);
+        }
+    }
+
+    // -----------------------------------------------------------------------
+    // Main dispatch  (mirrors Python's node_2_usj + dispatch_map)
+    // -----------------------------------------------------------------------
+
+    pub fn node_2_usj(&mut self, node: Node, parent: &mut Value) {
+        let raw_kind = node.kind();
+        let node_type = raw_kind.replace('\\', "");
+        let node_type = node_type.as_str();
+
+        match node_type {
+            // --- noops ---
+            "" | "|" | "usfm" => {}
+
+            // --- text ---
+            "text"      => self.push_text_node(node, parent),
+            "verseText" => self.handle_verse_text(node, parent),
+
+            // --- structural ---
+            "id"      => self.node_2_usj_id(node, parent),
+            "chapter" => self.node_2_usj_chapter(node, parent),
+            "v"       => self.node_2_usj_verse(node, parent),
+
+            // --- paragraph-level ---
+            "paragraph" => self.node_2_usj_para(node, parent),
+            "pi" | "ph"             => self.node_2_usj_para(node, parent),
+
+            // --- generic para-style ---
+            "cl" | "cp" | "vp"     => self.node_2_usj_generic(node, parent),
+
+            // --- alt numbering ---
+            "ca" | "va" => self.node_2_usj_ca_va(node, parent),
+
+            // --- table ---
+            "table" | "tr" => self.node_2_usj_table(node, parent),
+
+            // --- milestones ---
+            "milestone" | "zNameSpace" => self.node_2_usj_milestone(node, parent),
+
+            // --- specials ---
+            "esb" | "cat" | "fig" | "ref" => self.node_2_usj_special(node, parent),
+
+            other => {
+                // Note markers
+                if in_set(other, NOTE_MARKERS) {
+                    self.node_2_usj_notes(node, parent);
+                // Char style markers (including Nested variants) + xt_standalone
+                } else if in_set(other, CHAR_STYLE_MARKERS)
+                    || is_nested_char(other)
+                    || other == "xt_standalone"
+                {
+                    self.node_2_usj_char(node, parent);
+                // Table cells
+                } else if in_set(other, TABLE_CELL_MARKERS) {
+                    self.node_2_usj_table(node, parent);
+                // Para style markers
+                } else if in_set(other, PARA_STYLE_MARKERS) {
+                    self.node_2_usj_generic(node, parent);
+                // Attribute nodes
+                } else if other.ends_with("Attribute") {
+                    self.node_2_usj_attrib(node, parent);
+                // Unknown — recurse into children
+                } else if node.child_count() > 0 {
+                    let mut cursor = node.walk();
+                    let children: Vec<Node> = node.children(&mut cursor).collect();
+                    for child in children {
+                        self.node_2_usj(child, parent);
+                    }
+                }
+            }
+        }
     }
 }
