+++ rust-usfm-parser/tests/test_list_conversion.rs
